[#31]feat : 로그아웃 연결
