+++ package.json
@@ -19,12 +19,12 @@
     "cheerio": "^1.0.0-rc.12",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "cookies-next": "^4.2.1",
     "date-fns": "^3.6.0",
     "dayjs": "^1.11.11",
     "lodash": "^4.17.21",
     "lucide-react": "^0.400.0",
     "next": "14.2.4",
-    "nookies": "^2.5.2",
     "path-to-regexp": "^7.1.0",
     "react": "^18",
     "react-day-picker": "^8.10.1",
