+++ src/app/(rootLayout)/(backofficeLayout)/layout.tsx
@@ -1,6 +1,5 @@
 import BackOfficeMenuBar from '@/components/organisms/backOfficeMenuBar';
 import { PropsWithChildren } from 'react';
-
 const BackOfficeLayout = ({ children }: PropsWithChildren) => {
   return (
     <div className="flex min-h-[100vh]">
