+++ src/utils/clientAPI.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { parseCookies } from 'nookies';
+import { getCookie } from 'cookies-next';
 
 const BACKEND_URL = process.env.NEXT_PUBLIC_BACKEND_URL;
 
@@ -13,9 +13,7 @@ export const client = axios.create({
 
 client.interceptors.request.use(
   async (config) => {
-    const cookies = parseCookies();
-    const accessToken = cookies.accessToken;
-
+    const accessToken = getCookie('AT');
     if (accessToken) {
       config.headers.Authorization = `Bearer ${accessToken}`;
     } else {
@@ -30,33 +28,12 @@ client.interceptors.request.use(
 client.interceptors.response.use(
   (response) => {
     console.log('응답 받음:', response.status, response.config.url);
-
     return response;
   },
 
   async (error) => {
     console.error('API 오류:', error.response?.status, error.response?.data);
-    //   const originalRequest = error.config;
-    //   if (error.response.status === 401 && !originalRequest._retry) {
-    //     originalRequest._retry = true;
-    //     try {
-    //       const response = await client.post(
-    //         '/api/v1/users/reissue',
-    //         {},
-    //         { withCredentials: true },
-    //       );
-    //       const newAccessToken = response.data.accessToken;
-    //       setCookie(null, 'accessToken', newAccessToken, {
-    //         maxAge: 30 * 24 * 60 * 60,
-    //         path: '/',
-    //       });
-    //       originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
-    //       return client(originalRequest);
-    //     } catch (refreshError) {
-    //       //로그아웃 로직
-    //       return Promise.reject(refreshError);
-    //     }
-    //   }
-    //   return Promise.reject(error);
+    // 토큰 재발행 로직 추가 예정
+    return Promise.reject(error);
   },
 );
